Update README.md
I had created an image and booted it, and found ssh to be disabled, I found that I needed the ssh file in the boot partition, when I put it there and restarted the pi no ssh was available.  Reformatting the disk and reloading raspbian and then adding the ssh file and then booting gave me a headless pi with ssh enabled.  My change make it more clear the process.